+++ app/(home)/community/[id]/edit-image.tsx
@@ -1,0 +1,5 @@
+import { EditImageScreen } from 'appdeptus/modules/communities/screens'
+
+const EditImageRoute = () => <EditImageScreen />
+
+export default EditImageRoute