+++ modules/communities/api/endpoints/getCommunityList.ts
@@ -20,6 +20,7 @@ const getCommunityList = (builder: CoreEndpointBuilder<CommunitiesApiTags>) =>
                 name,
                 created_at,
                 secret,
+                image,
                 members: communities_users (
                   *,
                   user (