+++ notifications/api/hooks/useNotificationUpdates.ts
@@ -4,17 +4,21 @@ import { useAppDispatch } from 'appdeptus/store'
 import { mapNullToUndefined } from 'appdeptus/utils'
 import pushApi from '../api'
 import { notificationsSchema } from '../schema'
-import useGameUpdateListener from './useNotificationListener'
+import useNotificationUpdateListener from './useNotificationListener'
 
 const useNotificationUpdates = () => {
   const dispatch = useAppDispatch()
 
   const { data: user } = useGetUserProfileQuery()
 
-  useGameUpdateListener({
+  useNotificationUpdateListener({
     eventHandler: ({ new: newData }) => {
+      console.log(JSON.stringify(newData))
+
       if (newData.notifications.communities) {
-        dispatch(communitiesApi.util.invalidateTags(['requests']))
+        dispatch(
+          communitiesApi.util.invalidateTags(['requests', 'communities'])
+        )
       }
 
       dispatch(