+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 - User profile images 📸
 - Community profile images 🖼️
+- Links to user profiles 🔗
 
 ### Changed